*add doctors schedule table

+++ src/app/lists/timeslots-list/timeslots-list.component.ts
@@ -1,6 +1,6 @@
 import {Component, OnInit, ViewChild} from '@angular/core';
 import {CompositeService} from '../../services/composite-service.service';
-import {MatSort, MatTableDataSource, Sort} from '@angular/material';
+import {MatSort, MatTableDataSource} from '@angular/material';
 import {TimeslotService} from '../../services/timeslot-service.service';
 
 
@@ -13,11 +13,10 @@ export class TimeslotsListComponent implements OnInit {
 
   displayedColumns: string[] = ['date', 'time', 'specialty', 'doctor', 'cabinet', 'patient', 'delete'];
   dataSource;
+  @ViewChild(MatSort, {static: true}) sort: MatSort;
 
   constructor(private compositeService: CompositeService, private timeSlotService: TimeslotService) {
   }
-
-  @ViewChild(MatSort, {static: true}) sort: MatSort;
 
   ngOnInit() {
     this.reloadData();
@@ -28,22 +27,24 @@ export class TimeslotsListComponent implements OnInit {
   }
 
   reloadData() {
-    this.compositeService.findAll().subscribe(data => {
+    this.compositeService.getTimeslots().subscribe(data => {
       const sorted = data.sort((a, b) => a.time.localeCompare(b.time))
-                          .sort((a, b) => a.doctor.localeCompare(b.doctor))
-                          .sort((a, b) => a.date.localeCompare(b.date));
+      .sort((a, b) => a.doctor.localeCompare(b.doctor))
+      .sort((a, b) => a.date.localeCompare(b.date));
       this.dataSource = new MatTableDataSource(sorted);
     });
   }
 
   deleteTimeslot(id: number) {
-    this.timeSlotService.deleteTimeslot(id)
-    .subscribe(
-        data => {
-          console.log(data);
-          this.reloadData();
-        },
-        error => console.log(error));
+    if (confirm('Уверены что хотите удалить?')) {
+      this.timeSlotService.deleteTimeslot(id)
+      .subscribe(
+          data => {
+            console.log(data);
+            this.reloadData();
+          },
+          error => console.log(error));
+    }
   }
 }
